+++ chroma.go
@@ -1,0 +1,9 @@
+package glamour
+
+import "github.com/charmbracelet/glamour/ansi"
+
+func init() {
+	for _, style := range DefaultStyles {
+		ansi.ChromaRegister(style)
+	}
+}
